requirements.txt
# Core Frameworks
fastapi==0.120.4
uvicorn[standard]==0.38.0
pydantic===2.12.3
pydantic-settings==2.11.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
asyncpg==0.30.0
aiosqlite==0.21.0
greenlet==3.2.4
alembic==1.17.1

# Cache
redis[hiredis]==7.0.1

# Scheduled Tasks
apscheduler==3.10.4

# HTTP Client
httpx==0.28.1
requests==2.32.5

# GTFS & Transit Data
gtfs-realtime-bindings==1.0.0
protobuf==6.33.0

# Security & Authentication
python-jose[cryptography]==3.5.0
pwdlib[argon2]==0.3.0
python-multipart==0.0.20
email-validator==2.3.0

# Data Processing
pandas==2.3.3
numpy==2.3.4

# Data & Time
python-dateutil==2.9.0.post0
pytz==2025.2

# Environment Variables
python-dotenv==1.2.1

# Testing (TDD)
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1

# Development Tools
black==25.9.0
flake8==7.3.0
mypy==1.18.2